enable isolation of resource
